+++ ChangeLog.md
@@ -1,8 +1,11 @@
 # Change Log
 
 ## 2022-04-07
+### Enhanchements
 * Ensure user is properly transfered when merging people, including user email updated to match authority person.
 * Display system roles and additional permissions on Users List Screen.
+* Add All members export for Parent groups.
+
 ## 2022-04-05
 ### Fixes
 * Bring all activity log UI in line with optimized data coming from the server.